luci-app-ddns: check existence of nslookup util rather than making DNS request
This saves a DNS request whenever the check is performed.  For systems with nameservers configured in `/etc/resolv.conf`, nslookup will sent a request to those servers (eg bypassing /etc/hosts), thus taking a lot more time than necessary.  For servers that are remote, this will take a *lot* longer than just checking if the nslookup util exists.

This change follows in the same vein as all the other checks and just using `which` without actually invoking the util's functionality.

Signed-off-by: David Beitey <david@davidjb.com>